% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_flood_depths2.R
\name{calc_flood_depths2}
\alias{calc_flood_depths2}
\title{Calculate flooding depths and marsh flooding metrics from a user-supplied data frame or saved file of raw water level data}
\usage{
calc_flood_depths2(
  data,
  record = "water level",
  marsh_elev,
  print_results = TRUE,
  return_all = FALSE
)
}
\arguments{
\item{data}{data frame. A data frame of raw water level data.}

\item{record}{string (optional). The name of the data record to be used in the output data frame of flooding metrics. Defaults to "water level".}

\item{marsh_elev}{number. The elevation of the marsh in meters NAVD88.}

\item{print_results}{TRUE/FALSE. Do you want to print the calculated flood metrics to the console? Defaults to TRUE.}

\item{return_all}{TRUE/FALSE. Do you want to retun a list of dataframes including all raw data, intermediate data, and flood depths? Defaults to FALSE.}
}
\value{
A dataframe with flooding depths or list of data frames with all raw data, intermediate data and flooding depth data.
}
\description{
Calculate flooding depths and marsh flooding metrics from a user-supplied data frame or saved file of raw water level data
}
\note{
Adapted from \url{https://github.com/AndrewBirchHydro/albAquariusTools}.
}
\examples{
calc_flood_depths2(data = )
}
