% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_recent_local_slr2.R
\name{calc_recent_local_slr2}
\alias{calc_recent_local_slr2}
\title{Calculate the recent local rate of SLR from a user-supplied data frame or saved file of raw water level data}
\usage{
calc_recent_local_slr2(
  data,
  record = "water level",
  print_results = TRUE,
  return_all = FALSE
)
}
\arguments{
\item{data}{data frame. A data frame of raw water level data.}

\item{record}{string (optional). The name of the data record to be used in the output data frame of recent local slr. Defaults to "water level".}

\item{print_results}{TRUE/FALSE. Do you want to print the calculated recent local SLR rate to the console? Defaults to TRUE.}

\item{return_all}{TRUE/FALSE. Do you want to retun a list of dataframes including all raw data, intermediate data, and recent local slr rate? Defaults to FALSE.}

\item{marsh_elev}{number. The elevation of the marsh in meters NAVD88.}
}
\value{
A dataframe with the recent local slr rate or list of data frames with all raw data, intermediate data and recent local slr rate.
}
\description{
Calculate the recent local rate of SLR from a user-supplied data frame or saved file of raw water level data
}
\note{
Adapted from \url{https://github.com/AndrewBirchHydro/albAquariusTools}.
}
\examples{
calc_recent_local_slr2(data = )
}
