% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_tidal_datums1.R
\name{calc_tidal_datums1}
\alias{calc_tidal_datums1}
\title{Get raw water level data from Aquarius and calculate tidal datums}
\usage{
calc_tidal_datums1(
  Location,
  window = 11.5,
  start_date = NULL,
  end_date = NULL,
  print_results = TRUE,
  return_all = FALSE
)
}
\arguments{
\item{Location}{string. The name of the location in the Aquarius database.}

\item{window}{number. Size of the time window in hours used to calculate MHW and MLW. Defaults to 11.5 hours.}

\item{start_date}{string (optional). Filter time series to data after a
specific date.}

\item{end_date}{string (optional). Filter time series to data before a
specific date.}

\item{print_results}{TRUE/FALSE. Do you want to print the calculated tidal datums to the console? Defaults to TRUE.}

\item{return_all}{TRUE/FALSE. Do you want to retun a list of dataframes including all raw data, intermediate data, and tidal datums? Defaults to FALSE.}
}
\value{
A dataframe with tidal datums or list of data frames with all raw data, intermediate data and tidal data.
}
\description{
Get raw water level data from Aquarius and calculate tidal datums
}
\note{
Adapted from \url{https://github.com/AndrewBirchHydro/albAquariusTools}.
}
\examples{
calc_tidal_datums(Location = "COLO_M30_WaterLevel", start_date = "2024-01-01", end_date = "2024-12-31")

}
