% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_tidal_datums2.R
\name{calc_tidal_datums2}
\alias{calc_tidal_datums2}
\title{Calculate tidal datums from a user-supplied data frame or saved file of raw water level data}
\usage{
calc_tidal_datums2(
  data,
  Location = NULL,
  record = "water level",
  window = 11.5,
  print_results = TRUE,
  return_all = FALSE
)
}
\arguments{
\item{data}{data frame. A data frame of raw water level data.}

\item{Location}{string (optional). The name of the location to be used in the output data frame of tidal datums.}

\item{record}{string (optional). The name of the data record to be used in the output data frame of tidal datums. Defaults to "water level".}

\item{window}{number. Size of the time window in hours used to calculate MHW and MLW. Defaults to 11.5 hours.}

\item{print_results}{TRUE/FALSE. Do you want to print the calculated tidal datums to the console? Defaults to TRUE.}

\item{return_all}{TRUE/FALSE. Do you want to retun a list of dataframes including all raw data, intermediate data, and tidal datums? Defaults to FALSE.}
}
\value{
A dataframe with tidal datums or list of data frames with all raw data, intermediate data and tidal data.
}
\description{
Calculate tidal datums from a user-supplied data frame or saved file of raw water level data
}
\note{
Adapted from \url{https://github.com/AndrewBirchHydro/albAquariusTools}.
}
\examples{
calc_tidal_datums2(data = )
}
