% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/daily_mean_flood.R
\name{daily_mean_flood}
\alias{daily_mean_flood}
\title{Calculate the daily mean minimum, maximum, or average daily flood depth}
\usage{
daily_mean_flood(flood_depths_df, summary_metric = "mean")
}
\arguments{
\item{flood_depths_df}{(required); A data frame containing flood depth data
obtained via the function \code{flood_depths}. It must include columns for
'flood_depth', 'date', 'Identifier', 'Name', 'LocationIdentifier', and
'elev_navd88' - a column of marsh surface elevation values in meters NAVD88.
.}

\item{summary_metric}{character (required); The flood depth summary metric to
calculate. One of:
\itemize{
\item "min": the daily mean minimum flood depth above the marsh surface.
\item "max": the daily mean maximum flood depth above the marsh surface.
\item "average": the daily mean average flood depth above the marsh surface.
}}
}
\value{
A data frame with the daily mean minimum, maximum, or average flood
depth and flood elevation above the marsh surface at each site. Data frame
columns are defined as:
\itemize{
\item Name: The site name e.g., Assateague Island NS - M5 -Pope Bay Marsh.
\item Identifier: The site name and time series name e.g., Depth.meters NAVD88 ASIS_M5_Waterlevel.
\item LocationIdentifier: The time series name e.g., ASIS_M5_Waterlevel.
\item park: Four-letter park code.
\item site_name: The SET site name e.g., Marsh 5 (Pope Bay).
\item daily_mean_min/max/avg_flood_depth: The daily mean minimum, maximum, or average depth of flooding above the marsh surface in meters.
\item elev_navd88: The elevation of the marsh surface in meters NAVD88.
\item daily_mean_min/max/avg: The daily mean minimum, maximum, or average water level in meters NAVD88
}
}
\description{
This function calculates mean minimum, maximum, or average daily flood depth
from a flood depth dataset obtained via the function \code{flood_depths}. The
mean daily flood depth is an indicator of marsh drainage and is calculated as
the average difference between the daily minimum, maximum, or average water
elevation and the marsh elevation at each site.
}
\examples{
asis_daily_mean_min_flood <- daily_mean_flood(flood_depths_df = asis_flood_depths, summary_metric = "min")

asis_daily_mean_avg_flood <- daily_mean_flood(asis_flood_depths, summary_metric = "average")

}
\seealso{
\code{\link[=flood_depths]{flood_depths()}}
}
