% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_extremes.R
\name{find_extremes}
\alias{find_extremes}
\title{Find local extremes (peaks or troughs) in a numeric vector within a rolling
window}
\usage{
find_extremes(x, fun, window = 11.5)
}
\arguments{
\item{x}{(required); A numeric vector in which to find the extremes.}

\item{fun}{string (required); The function to apply within the rolling
window, either \code{max} for peaks or \code{min} for troughs.}

\item{window}{string (required); The approximate width of the window to use
for comparison. The actual number of elements in the window is calculated
as \code{(window * 4)/2}, rounded down. Default window is 11.5 for a 24 hour
day.}
}
\value{
A numeric vector of the same length as \code{x} where extreme values are
preserved and all other values are set to \code{NA_real_}.
}
\description{
This function identifies local maximums or minimums in a vector \code{x} by
comparing each element to a rolling window of its neighbors. It uses
\code{zoo::rollapply} for the initial rolling calculation and ensures that the
identified extreme value is unique within its local window to avoid marking
entire flat sections as extremes.
}
\examples{
# Find high and low extremes
df_extremes <- asis_wl \%>\%
  group_by(Identifier, Name, LocationIdentifier, .add = TRUE) \%>\%
  nest() \%>\%
  mutate(highs = map(data, ~find_extremes(.x$water_level, max)),
         lows = map(data, ~find_extremes(.x$water_level, min))) \%>\%
  unnest(cols = c(data, highs, lows)) \%>\%
  mutate(date_chr = lubridate::as_date(date))
  
}
