% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flood_depths.R
\name{flood_depths}
\alias{flood_depths}
\title{Calculate flooding depth above the marsh surface}
\usage{
flood_depths(wl_data, elev_df)
}
\arguments{
\item{wl_data}{(required); A data frame containing water level data obtained
via the function \code{get_wl_data}. It must include columns for 'water_level',
'date', 'Identifier', 'Name', and 'LocationIdentifier'.}

\item{elev_df}{(required); A data frame containing the marsh surface
elevation values in meters NAVD88 for each site. Must also contain either
an 'Identifier', 'Name', or 'LocationIdentifier' column that matches the
sites of interest in the \code{wl_data} data frame.}
}
\value{
A data frame showing the water level above the marsh surface at each
time step. Data frame columns are defined as:
\itemize{
\item Name: The site name e.g., Assateague Island NS - M5 -Pope Bay Marsh.
\item Identifier: The site name and time series name e.g., Depth.meters NAVD88 ASIS_M5_Waterlevel.
\item LocationIdentifier: The time series name e.g., ASIS_M5_Waterlevel.
\item park: Four-letter park code.
\item Unit: The water level measurement unit e.g. m or meters. Note the measurement units of the output will be the same as whatever was entered into Aquarius when the data was uploaded (either raw meters or meters referenced to NAVD88).
\item Label: The water level unit label e.g., meters NAVD88.
\item datetime: The Posix date/time of the water level measurement (YYYY-MM-DD H:M:S)
\item date: ISO date of the water level  measurement (YYYY-MM-DD).
\item time: The measurement time in military time format (H:M:S).
\item water_level: The water level depth in meters; this is either referenced to NAVD88 or, if not referenced to NAVD88, represents a water level depth relative to the vertical position of the water logger.
\item is_navd88: True/False - is the water level depth referenced to NAVD88?
\item site_name: The SET site name e.g., Marsh 5 (Pope Bay).
\item elev_navd88: The elevation of the marsh surface in meters NAVD88.
\item flood_depth: The depth of the water relative to the marsh surface in meters. Negative values indicate the water level was below the marsh surface and positive values indicate the water level was above the marsh surface.
\item wl_above_marsh: The water level in meters NAVD88 - Note that time steps where the water level was below the marsh surface are coded as NA's.
}
}
\description{
This function calculates the water depth above the marsh surface at each time
step for a water level dataset obtained via the function \code{get_wl_data} and a
user supplied data frame of marsh surface elevations for each site referenced
to meters NAVD88.
}
\note{
This comparison will only work properly if both the water level data
and marsh elevation data are referenced to the same datum e.g. meters
NAVD88.
}
\examples{
# data frame of marsh surface elevations in meters NAVD88
asis_elev_df <- data.frame(LocationIdentifier = c("ASIS_M5_WaterLevel", "ASIS_M6_WaterLevel", "ASIS_M8_WaterLevel"), elev_navd88 = c(0.145, 0.1325, 0.200))

asis_flood_depths <- flood_depths(wl_data = asis_wl, elev_df = asis_elev_df)

}
\seealso{
\code{\link[=get_wl_data]{get_wl_data()}}
}
