% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flood_events.R
\name{flood_events}
\alias{flood_events}
\title{Calculate the number of times that the marsh was flooded.}
\usage{
flood_events(flood_depths_df)
}
\arguments{
\item{flood_depths_df}{(required); A data frame containing flood depth data
obtained via the function \code{flood_depths}. It must include columns for
'flood_depth', 'date', 'Identifier', 'Name', 'LocationIdentifier', and
'wl_above_marsh' - a column of the water level above the marsh at each time
step.}
}
\value{
A data frame with 1 row per site containing the total number of events where the water level over topped the marsh surface over the period of
record. Data frame columns are defined as:
\itemize{
\item Name: The site name e.g., Assateague Island NS - M5 -Pope Bay Marsh.
\item Identifier: The site name and time series name e.g., Depth.meters NAVD88 ASIS_M5_Waterlevel.
\item LocationIdentifier: The time series name e.g., ASIS_M5_Waterlevel.
\item park: Four-letter park code.
\item site_name: The SET site name e.g., Marsh 5 (Pope Bay).
\item total_flood_events: The number of events where the water level over topped the marsh surface.
}
}
\description{
This function calculates the events where the water level switched from below the marsh to above the marsh surface on successive timesteps.
}
\examples{
asis_flood_events <- flood_events(flood_depths_df = asis_flood_depths)

}
