% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_parameters.R
\name{get_parameters}
\alias{get_parameters}
\title{Print a list of the parameters in a data frame of downloaded water data}
\usage{
get_parameters(wl_data)
}
\arguments{
\item{wl_data}{(required); A data frame containing water level data obtained
via the function \code{get_wl_data}. It must include columns for 'water_level',
'date', 'Identifier', 'Name', and 'LocationIdentifier'.}
}
\value{
Prints a list of water parameters in the data to the console.
}
\description{
Print a list of the parameters in a data frame of downloaded water data
}
\examples{
get_parameters(wl_data = caco_dh_seagrass_wl)

}
