% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sites.R
\name{get_sites}
\alias{get_sites}
\title{Print a list of the sites in a data frame of downloaded water data}
\usage{
get_sites(wl_data)
}
\arguments{
\item{wl_data}{(required); A data frame containing water level data obtained
via the function \code{get_wl_data}. It must include columns for 'water_level',
'date', 'Identifier', 'Name', and 'LocationIdentifier'.}
}
\value{
Prints a list of sites in the data to the console.
}
\description{
Print a list of the sites in a data frame of downloaded water data
}
\examples{
get_sites(wl_data = asis_set_wl)

}
