% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_timeseries.R
\name{get_timeseries}
\alias{get_timeseries}
\alias{Get_timeseries1}
\alias{Get_timeseries2}
\title{Download Aquarius time series data}
\usage{
Get_timeseries1(record, start_date, end_date)

Get_timeseries2(record)
}
\arguments{
\item{record}{string. The concatenated name of a time series record + a
location name - e.g., "Depth.meters NAVD88@ASIS_M5_WaterLevel".}

\item{start_date}{string (optional). Filter time series to data after a
specific date.}

\item{end_date}{string (optional). Filter time series to data before a
specific date.}
}
\value{
A list of objects containing raw time series water lavel data and
other metadata including:
\itemize{
\item TimeSeries - data record IDs, location IDs, parameters, units.
\item TimeRange - the start and end time of the data record.
\item NumPoints - the number of observations in the record.
\item \strong{Points} - the raw time series water level data. *This is the
data used by all other functions in the package.
\item ResponseVersion
\item ResponseTime - elapsed time for the API call.
\item Summary - info source description and disclaimer.
}
}
\description{
Download Aquarius time series data
}
\details{
The object "Points" from the output is the data that is used to
calculate tidal datums and is used by all other functions in the package.
}
\note{
Adapted from
\url{https://github.com/AndrewBirchHydro/albAquariusTools}.
}
\examples{
## global variables for api connection - only run once per session
timeseries = timeseriesClient()
timeseries$connect("https://aquarius.nps.gov/aquarius", "aqreadonly", "aqreadonly")
publishapiurl = 'https://aquarius.nps.gov/aquarius/Publish/v2'

# Get water level data for COLO M30 for 2024 
Get_timeseries1(record = "Depth.Meters NAVD88@COLO_M30_WaterLevel", start_date = "2024-01-01", end_date = "2024-12-31")

# Get water level data for COLO M30 entire record
Get_timeseries(record = "Depth.Meters NAVD88@COLO_M30_WaterLevel")

# Extract the data frame of raw water level data 
Get_timeseries1(record = "Depth.Meters NAVD88@COLO_M30_WaterLevel", start_date = "2024-01-01", end_date = "2024-12-31") \%>\%
  purrr::pluck("Points")

}
