% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wl_data.R
\name{get_wl_data}
\alias{get_wl_data}
\title{Load water level data from NPS Aquarius for a specific park}
\usage{
get_wl_data(park_code)
}
\arguments{
\item{park_code}{string (required); The 4 character park code - must be
capitalized (e.g., "ASIS", "GATE"). Options include ASIS, ACAD, BISC, CACO,
COLO, FIIS, GATE, GWMP, NACE, SARI, or VIIS.}
}
\value{
A data frame of water level (i.e., water level) data from the
selected park. Data frame columns are defined as:
\itemize{
\item Name: The site name e.g., Assateague Island NS - M5 -Pope Bay Marsh.
\item Identifier: The site name and time series name e.g., Depth.meters NAVD88 ASIS_M5_Waterlevel.
\item LocationIdentifier: The time series name e.g., ASIS_M5_Waterlevel.
\item park: Four-letter park code.
\item Unit: The water level measurement unit e.g. m or meters. Note the measurement units of the output will be the same as whatever was entered into Aquarius when the data was uploaded (either raw meters or meters referenced to NAVD88).
\item Label: The water level unit label e.g., meters NAVD88.
\item datetime: The Posix date/time of the water level measurement (YYYY-MM-DD H:M:S)
\item date: ISO date of the water level  measurement (YYYY-MM-DD).
\item time: The measurement time in military time format (H:M:S).
\item water_level: The water level depth in meters; this is either referenced to NAVD88 or, if not referenced to NAVD88, represents a water level depth relative to the vertical position of the water logger.
\item is_navd88: True/False - is the water level depth referenced to NAVD88?
\item site_name: The SET site name e.g., Marsh 5 (Pope Bay).
}
}
\description{
Utilizes functions from the package
\link[=https://github.com/nationalparkservice/imd-fetchaquarius]{imd-fetchaquarius}
to connect to the NPS Aquarius database.
}
\note{
This function gets water level data for sites that have either "Depth"
or "Water Level" as one of the parameters. Output units will be the same as
whatever was entered into Aquarius when the data was uploaded (either raw
meters or meters referenced to NAVD88).
}
\examples{
library(tidyverse)
library(remotes)
library(fetchaquarius)

asis_wl <- get_wl_data(park = "ASIS")

}
