% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_raw_wl.R
\name{load_raw_wl}
\alias{load_raw_wl}
\title{Load raw water level data for a single site from a csv, xlsx, or xls file}
\usage{
load_raw_wl(
  file_path = NULL,
  time_col = "Timestamp",
  value_col = "NumericValue1"
)
}
\arguments{
\item{file_path}{string. Either the full path to the saved file or a relative path specified with \code{\link[here]{here}}.}

\item{time_col}{string. The name of the column with date-times. Defaults to "Timestamp" based on the output of \code{\link{Get_timeseries1}} or \code{\link{Get_timeseries2}}. Required if \code{data} has date-times in a column other than "Timestamp".}

\item{value_col}{string. The name of the column containing the water level values. Defaults to "NumericValue1" based on the output of \code{\link{Get_timeseries1}} or \code{\link{Get_timeseries2}}. Required if \code{data} has water level values in a column other than "NumericValue1".}
}
\value{
a data frame with time series water level data.
}
\description{
Load raw water level data for a single site from a csv, xlsx, or xls file
}
\examples{
data <- load_raw_wl(here::here("data", "ncbn_aq_points.csv"), time_col = "time", value_col = "wl")

}
