% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longest_flood.R
\name{longest_flood}
\alias{longest_flood}
\title{Calculate the longest time that the marsh stayed flooded.}
\usage{
longest_flood(flood_depths_df)
}
\arguments{
\item{flood_depths_df}{(required); A data frame containing flood depth data
obtained via the function \code{flood_depths}. It must include columns for
'flood_depth', 'date', 'Identifier', 'Name', 'LocationIdentifier', and
'wl_above_marsh' - a column of the water level above the marsh at each time
step.}
}
\value{
A data frame containing the longest flood event at each site. Data
frame columns are defined as:
\itemize{
\item Name: The site name e.g., Assateague Island NS - M5 -Pope Bay Marsh.
\item Identifier: The site name and time series name e.g., Depth.meters NAVD88 ASIS_M5_Waterlevel.
\item LocationIdentifier: The time series name e.g., ASIS_M5_Waterlevel.
\item park: Four-letter park code.
\item site_name: The SET site name e.g., Marsh 5 (Pope Bay).
\item consecutive_group: The flood event group ID for the longest flood event.
\item data: A nested data of all flood events at each site.
\item consecutive_count: The total number of time steps recorded during the longest flood event.
\item min_date_longest_flood: The date that the longest consecutive flood event started.
\item max_date_longest_flood: The date that the longest consecutive flood event ended.
\item max_consecutive_flood_days: The number of days that the marsh was flooded during the longest consecutive flood event.
}
}
\description{
This function finds the longest consecutive time that the water level
remained above the marsh surface.
}
\examples{
asis_longest_flood <- longest_flood(asis_flood_depths)

}
