% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_water_level_with_datums.R
\name{plot_water_level_with_datums}
\alias{plot_water_level_with_datums}
\title{Plot NAVD88 water level data with tidal datums}
\usage{
plot_water_level_with_datums(
  wl_data,
  site = NA,
  x_right_expand = 0.25,
  include_extremes = FALSE
)
}
\arguments{
\item{wl_data}{(required); A data frame containing water level data obtained
via the function \code{get_wl_data}. It must include columns for 'water_level',
'date', 'Identifier', 'Name', and 'LocationIdentifier'.}

\item{site}{(optional): A string specifying the name of a single site to
plot.}

\item{x_right_expand}{(optional): Value used to expand the x-axis to the
right.}

\item{include_extremes}{(optional): TRUE will plot MHHW and MLLW in addition
to MHW and MLW. Defaults to FALSE.}
}
\value{
A single ggplot.
}
\description{
This function uses the tidal_datums() function to calculate tidal datums and
then plot them on top of the water level data referenced to NAVD88.
}
\examples{
plot_water_level_with_datums(wl_data = asis_m8_set_wl)

}
