% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_aquarius_data.R
\name{query_aquarius}
\alias{query_aquarius}
\alias{Print_Locations}
\alias{Print_datasets}
\alias{Print_metadata}
\title{Query info about Aquarius locations and datasets}
\usage{
Print_Locations(Network)

Print_datasets(Location)

Print_metadata(Location)
}
\arguments{
\item{Network}{string. The full name of an NPS I&M network e.g. "Northeast
Coastal and Barrier Network".}

\item{Location}{string. A site name from the Aquarius database e.g.
"ASIS_M5_WaterLevel".}
}
\description{
Query the Aquarius database to see available locations,
datasets, and metadata.
\itemize{
\item \code{Print_Locations()} shows all locations with available data in a
specific I&M network.
\item \code{Print_datasets()} shows all data sets available for a specific
location.
\item \code{Print_metadata()} shows metadata for a specific location.
}
}
\note{
Adapted from
\url{https://github.com/AndrewBirchHydro/albAquariusTools}.
}
\examples{
## global variables for api connection - only run once per session
timeseries = timeseriesClient()
timeseries$connect("https://aquarius.nps.gov/aquarius", "aqreadonly", "aqreadonly")
publishapiurl = 'https://aquarius.nps.gov/aquarius/Publish/v2'

Print_Locations(Network = "Northeast Coastal and Barrier Network")

Print_datasets(Location = "COLO_M30_WaterLevel")

Print_metadata(Location = "COLO_M30_WaterLevel")

## can also save results to a list or dataframe
ncbn_locations <- Print_Locations(Network = "Northeast Coastal and Barrier Network")

COLO_M30_metadata <- Print_metadata(Location = "COLO_M30_WaterLevel")

}
