% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidal_datums.R
\name{tidal_datums}
\alias{tidal_datums}
\title{Calculate tidal datums (MHW, MHHW, MLW, MLLW) from water level data}
\usage{
tidal_datums(wl_data)
}
\arguments{
\item{wl_data}{(required); A data frame containing water level data obtained
via the function \code{get_wl_data}. It must include columns for 'water_level',
'date', 'Identifier', 'Name', and 'LocationIdentifier'.}
}
\value{
A data frame containing the calculated tidal datums for each site.
Data frame columns are defined as:
\itemize{
\item Name: The site name e.g., Assateague Island NS - M5 -Pope Bay Marsh.
\item Identifier: The site name and time series name e.g., Depth.meters NAVD88 ASIS_M5_Waterlevel.
\item LocationIdentifier: The time series name e.g., ASIS_M5_Waterlevel.
\item park: Four-letter park code.
\item site_name: The SET site name e.g., Marsh 5 (Pope Bay).
\item MHW: Mean high water - Note the measurement units of the output will be the same as whatever was entered into Aquarius when the data was uploaded (either raw meters or meters referenced to NAVD88).
\item MHHW: Mean higher high water.
\item MLW: Mean low water.
\item MLLW: Mean lower low water.
\item first_date: The first date in the water level record.
\item last_date: The last date in the water level record.
\item record_count: The number of observations in the water level record.
}
}
\description{
This function calculates tidal datums from a water level dataset obtained via
the function \code{get_wl_data}. It groups the data to calculate datums for each
unique site. Internally, it relies on the function \code{find_extremes} to
identify daily high and low water levels.
}
\note{
Output units will be the same as whatever was entered into Aquarius
when the data was uploaded (either raw meters or meters referenced to
NAVD88).
}
\examples{
asis_tidal_datums <- tidal_datums(asis_wl)

}
\seealso{
\code{\link[=get_wl_data]{get_wl_data()}}, \code{\link[=find_extremes]{find_extremes()}}
}
