% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/time_flooded.R
\name{time_flooded}
\alias{time_flooded}
\title{Calculate the percent of time that the water level is above the marsh surface
over the course of the period of record.}
\usage{
time_flooded(flood_depths_df)
}
\arguments{
\item{flood_depths_df}{(required); A data frame containing flood depth data
obtained via the function \code{flood_depths}. It must include columns for
'flood_depth', 'date', 'Identifier', 'Name', 'LocationIdentifier', and
'wl_above_marsh' - a column of the water level above the marsh at each time
step.}
}
\value{
A data frame with 1 row per site containing the total time recorded,
the total time that the marsh was flooded over the period of record, and
the percentage of time that the marsh was flooded over the period of
record. Data frame columns are defined as:
\itemize{
\item Name: The site name e.g., Assateague Island NS - M5 -Pope Bay Marsh.
\item Identifier: The site name and time series name e.g., Depth.meters NAVD88 ASIS_M5_Waterlevel.
\item LocationIdentifier: The time series name e.g., ASIS_M5_Waterlevel.
\item park: Four-letter park code.
\item site_name: The SET site name e.g., Marsh 5 (Pope Bay).
\item total_time_recorded: The total time recorded.
\item flooding_time: The total time that the marsh was flooded over the period of record in hours.
\item percent_time_flooded: The percentage time that the marsh was flooded over the period of record.
}
}
\description{
This function calculates 1) the total time that the marsh was flooded over
the period of record, and 2) the percentage of time that the marsh was
flooded over the period of record.
}
\examples{
asis_time_flooded <- time_flooded(asis_flood_depths)

}
