% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TS_simplify.R
\name{TS_simplify}
\alias{TS_simplify}
\title{Simplify time series water level data}
\usage{
TS_simplify(data, time_col = "Timestamp", value_col = "NumericValue1")
}
\arguments{
\item{data}{can be either
\itemize{
\item A list of data frames produced by either or \code{\link{Get_timeseries1}} or \code{\link{Get_timeseries2}}. The list must contain a data frame with the name "Points"
that contains the raw time series data.
\item A single data frame of raw time series water level data. This data
frame can be produced by either 1) loading a saved file with the function
or \code{\link{load_raw_wl}} or 2) extracting the "Points" data frame from the output of
either of the \code{Get_timeseries} functions.
}}

\item{time_col}{string (optional). The name of the column containing the
date-times. Defaults to "Timestamp" based on the output of the
\code{Get_timeseries} functions. Required if \code{data} has date-times in a column
other than "Timestamp".}

\item{value_col}{string (optional). The name of the column containing the
water level values. Defaults to "NumericValue1" based on the output of the
\code{Get_timeseries} functions. Required if \code{data} has water level values in a
column other than "NumericValue1".}
}
\value{
A data frame with two columns: \code{date_time} which contains date-times
transformed into a usable format and \code{value} which contains the water level
value.
}
\description{
This function simplifies time series water level data into 2
columns with properly formatted date-time and water level columns.
}
\note{
Adapted from
\url{https://github.com/AndrewBirchHydro/albAquariusTools}.
}
\examples{
asis_2024_aq_data1 <- readRDS(here::here("data", "asis_2024_aq_data.Rds"))
asis_2024_aq_data_simple1 <- TS_simplify(asis_2024_aq_data1)

asis_2024_aq_data2 <- asis_2024_aq_data1 \%>\%
  purrr::pluck("Points")
asis_2024_aq_data_simple2 <- TS_simplify(asis_2024_aq_data2)   

asis_2024_aq_points1 <- readRDS(here::here("data", "asis_2024_aq_points.Rds")) 
asis_2024_aq_points_simple1 <- TS_simplify(asis_2024_aq_points1, time_col = "time", value_col = "wl")

asis_2024_aq_points2 <- load_raw_wl(file_path = here::here("data", "asis_2024_aq_points.csv"), time_col = "time", value_col = "wl")
asis_2024_aq_points_simple2 <- TS_simplify(asis_2024_aq_points2)

}
