% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_water_level_with_datums.R
\name{save_water_level_with_datums}
\alias{save_water_level_with_datums}
\title{Save NAVD88 water level data, tidal datums, and plot to an excel file}
\usage{
save_water_level_with_datums(wl_data, folder, file_name, overwrite = FALSE)
}
\arguments{
\item{wl_data}{(required); A data frame containing water level data obtained
via the function \code{get_wl_data}. It must include columns for 'water_level',
'date', 'Identifier', 'Name', and 'LocationIdentifier'. Note that the data
needs to be filtered to single site before using this function.}

\item{folder}{(required); A string with the folder that you want to save to.
Defaults to the current working directory unless you specify a full folder
path (e.g., "C:/Users/lfeher/data").}

\item{file_name}{(required); A string with the name of the excel file that
you want to use.}

\item{overwrite}{(optional); Do you want to overwrite an existing file of the
same name? Defaults to FALSE.}
}
\value{
Invisibly saves an excel workbook to the specified folder/file_name.
}
\description{
Save NAVD88 water level data, tidal datums, and plot to an excel file
}
\examples{
save_water_level_with_datums(wl_data = asis_m8_set_wl, folder = "data", file_name = "ASIS_M8_Plotting", overwrite = TRUE)
  
}
